+++ Wrappers/InvoiceWrapper.cs
@@ -108,5 +108,28 @@ namespace Facturapi.Wrappers
         {
             return this.DownloadAsync(id, "xml");
         }
+
+        private async Task<Stream> DownloadCancellationReceiptAsync(string id, string format)
+        {
+            var response = await client.GetAsync(Router.DownloadCancellationReceipt(id, format));
+            if (!response.IsSuccessStatusCode)
+            {
+                var resultString = await response.Content.ReadAsStringAsync();
+                var error = JsonConvert.DeserializeObject<JObject>(resultString, this.jsonSettings);
+                throw new FacturapiException(error["message"].ToString());
+            }
+            var stream = await response.Content.ReadAsStreamAsync();
+            return stream;
+        }
+
+        public Task<Stream> DownloadCancellationReceiptXmlAsync(string id)
+        {
+            return DownloadCancellationReceiptAsync(id, "xml");
+        }
+
+        public Task<Stream> DownloadCancellationReceiptPdfAsync(string id)
+        {
+            return DownloadCancellationReceiptAsync(id, "pdf");
+        }
     }
 }